submodule: port submodule subcommand 'summary' from shell to C
Convert submodule subcommand 'summary' to a builtin and call it via
'git-submodule.sh'.

The shell version had to call $diff_cmd twice, once to find the modified
modules cared by the user and then again, with that list of modules
to do various operations for computing the summary of those modules.
On the other hand, the C version does not need a second call to
$diff_cmd since it reuses the module list from the first call to do the
aforementioned tasks.

In the C version, we use the combination of setting a child process'
working directory to the submodule path and then calling
'prepare_submodule_repo_env()' which also sets the 'GIT_DIR' to '.git',
so that we can be certain that those spawned processes will not access
the superproject's ODB by mistake.

A behavioural difference between the C and the shell version is that the
shell version outputs two line feeds after the 'git log' output when run
outside of the tests while the C version outputs one line feed in any
case. This is because of the difference in their 'git log' calls: the
shell version does a '--pretty=format:  %m %s' while the C version does
'--pretty=  %m %s'. The latter is equivalent to 'tformat: %m %s'
therefore prints one terminator character (i.e., a newline) and is
followed by a 'printf("\n")' at the end which causes one line feed
whereas the shell version has two 'echo' calls at the end due to which
there are two newlines.

Also, when one tries to pass an unrecognised argument after a recognised
argument, we do not get an output since the argument, being treated as a
non-option argument, when passed on to $diff_cmd, results in nothing
being fetched by $diff_cmd as it does not have a valid output for that
path. This behvaiour is consistent for the shell and C version.

The error message in case of computing a summary for non-existent
submodules in the C version is different from that of the shell version.
Since the new error message is not marked for translation, change the
'test_i18ngrep' in t7421.4 to 'grep'.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Stefan Beller <stefanbeller@gmail.com>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Helped-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>