submodule: port submodule subcommand 'summary' from shell to C
Convert submodule subcommand 'summary' to a builtin and call it via
'git-submodule.sh'.

The shell version had to call $diff_cmd twice, once to find the modified
modules cared by the user and then again, with that list of modules
to do various operations for computing the summary of those modules.
On the other hand, the C version does not need a second call to
$diff_cmd since it reuses the module list from the first call to do the
aforementioned tasks.

In the C version, we use the combination of setting a child process'
working directory to the submodule path and then calling
'prepare_submodule_repo_env()' which also sets the 'GIT_DIR' to '.git',
so that we can be certain that those spawned processes will not access
the superproject's ODB by mistake.

Also, the shell version did not support the 'quiet' though it gives a
quietened output for one particular case due to a small loophole in the
shell implementation.

    git submodule summary HEAD --quiet

The above statement prints nothing due to this part of the shell script
where the $diff_cmd is called:

    modules=$(git $diff_cmd $cached --ignore-submodules=dirty --raw $head -- "$@" |
	    sane_egrep '^:([0-7]* )?160000' |
	    while read -r mod_src mod_dst sha1_src sha1_dst status sm_path
	    do
             .....

The "$@" contains the 'quiet' option, which is then passed to $diff_cmd
which in turn resulted in us getting no modules to compute a summary of.
The parser of 'cmd_summary()' does not question the 'quiet' after an
argument since it sees the 'HEAD' (or any specified revision) and stops
parsing the options thus never seeing the '--quiet' present after it.
The C version however does not work this way and if we pass an
unrecognized option (in this case 'quiet') after a non-optional argument,
the 'parse_options()' function takes timely action and errors out right
away, not giving a quietened output like we got in the shell version
thus, outputting the usage from 'submodule--helper.c':

error: unknown option `quiet'
usage: git submodule--helper summary [<options>] [commit] [--] [<path>]

    --cached              use the commit stored in the index instead of the submodule HEAD
    --files               to compare the commit in the index with that in the submodule HEAD
    --for-status          skip submodules with 'ignore_config' value set to 'all'
    -n, --summary-limit <n>
                          limit the summary size

This is one behavioural difference between the shell and the C version.
Though, one thing to keep in mind is that if we pass the unrecognized
option before the non-optional argument, the usage from the shell
version of 'summary' is printed.

The error message in case of computing a summary for non-existent
submodules in the C version is different from that of the shell version.
Since the new error message is not marked for translation, change the
'test_i18ngrep' in t7421.4 to 'grep'.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Stefan Beller <stefanbeller@gmail.com>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Helped-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>