submodule: port submodule subcommand 'summary' from shell to C
Convert submodule subcommand 'summary' to a builtin and call it via
'git-submodule.sh'.

The shell version of submodule subcommand 'summary' takes multiple
passes to compute the summary of the submodule(s) in a superproject
whereas the C version of 'summary' does the job in a single pass since
it can gather all the information in one go. In the C version, we use
the combination of setting a child process' working directory to the
submodule path and then calling 'prepare_submodule_repo_env()' which
also sets the 'GIT_DIR' to '.git', so that we can be certain that those
spawned processes will not access the superproject's ODB by mistake.
To add on, the shell version calls $diff_cmd twice, once to get the
modified modules cared by the user and then again, with that list of
modules; the C version does not need a separate $diff_cmd call to
find the modified modules since the module list is fetched by
'submodule_summary_callback()'.
Also, the shell version has two newlines after the log message while
the C version has just one newline.

The option parsing for the 'quiet' option was missing in the remnants
of the shell script version of 'summary' due to which the option was
no recognised when passed by the user. Therefore, implement parsing
for the 'quiet' option.

While at it, eliminate the 'get_submodule_config()' function from
the shell script since it was used only by 'cmd_summary()'. Change the
'test_i18ngrep' in t7421.4 to 'grep' since the error message prompted
in the test is not marked for translation.

-----8<-----
The submodule subcommand 'summary' is ported in the process of
making git-submodule a builtin. The function cmd_summary() from
git-submodule.sh is ported to functions module_summary(),
compute_summary_module_list(), prepare_submodule_summary() and
generate_submodule_summary(), print_submodule_summary().

The first function module_summary() parses the options of submodule
subcommand and also acts as the front-end of this subcommand.
After parsing them, it calls the compute_summary_module_list()

The functions compute_summary_module_list() runs the diff_cmd,
and generates the modules list, as required by the subcommand.
The generation of this module list is done by the using the
callback function submodule_summary_callback(), and stored in the
structure module_cb.

Once the module list is generated, prepare_submodule_summary()
further goes through the list and filters the list, for
eventually calling the generate_submodule_summary() function.

The function generate_submodule_summary() takes care of generating
the summary for each submodule and then calls the function
print_submodule_summary() for printing it.
----->8------
Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Stefan Beller <stefanbeller@gmail.com>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Helped-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>