submodule: port submodule subcommand 'summary' from shell to C
Convert submodule subcommand 'summary' to a builtin and call it via
'git-submodule.sh'.

The shell version had to call $diff_cmd twice, once to find the modified
modules cared by the user and then again, to find the error message (if
any), 'sha1_abbr_{src,dst}' and find the range of commits to compute the
summary of. On the other hand, the C version does not need a second
call to $diff_cmd since it reuses the module list from the first call
to do the aforementioned tasks.

In the C version, we use the combination of setting a child process'
working directory to the submodule path and then calling
'prepare_submodule_repo_env()' which also sets the 'GIT_DIR' to '.git',
so that we can be certain that those spawned processes will not access
the superproject's ODB by mistake.

Also, the shell version did not support the 'quiet' though it gives a
quitened output for one particular case due to a small loophole in the
shell implementation.

    git submodule summary HEAD --quiet

When a 'quiet' comes after any argument of 'git submodule summary', we
get a quietened output due to this part of the shell script where the
$diff_cmd is called:

    modules=$(git $diff_cmd $cached --ignore-submodules=dirty --raw $head -- "$@" |

The "$@" contains the 'quiet' option therefore the output is quitened.
The parser of 'cmd_summary()' does not question the 'quiet' after an
argument since it assumes it to be another argument we supplied. The C
version however does not work this way and parses out the 'quiet' option
no matter where it is placed in the command due to the 'parse_options()'
function called in 'module_summary()'. This is one behavioural
difference between the shell and the C version.

The error message in case of computing a summary for non-existent
submodules in the C version is different from that of the shell version.
Since the new error message is not marked for translation, change the
'test_i18ngrep' in t7421.4 to 'grep'.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Stefan Beller <stefanbeller@gmail.com>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Helped-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>