submodule: fix failure of t7418
Test 't7418-submodule-sparse-gitmodules.sh' fails its test 6 which tests
the functionality of the 'for-status' option of 'git-submodule summary'.
Neutralize the failure point by calling the function
'is_non_bare_repository()' instead of 'is_git_directory()' since there
is a chance that there only exists a gitfile to represent the submodule
instead of a gitdir.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oO@gmail.com>