submodule: rename helper functions for more clarity
Rename the helper functions, 'show_submodule_summary();,
'prepare_submodule_summary()' and 'print_submodule_summary()' defined in
submodule.c to '*_submodule_diff_summary()'. These functions are called
by 'builtin_diff()' in diff.c. The reason for this change is that
functions of the same name will be used by the C port of 'git submodule
summary' in the following commit(s) hence the function names may cause
confusion if not changed.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>