Compare with empty tree
Dscho says:
It does not actually read from `stdin`. It reads from `/dev/null`,
redirected to the input. And what it _actually_ does is to generate the
OID of the empty tree.

But we already _have_ the OID of the empty tree! It's
`the_hash_algo->empty_tree`.
-------------------------------------------
Therefore pass Hash of empty tree directly hence eliminating the checks
placed by fstat() and index_fd(). The check was:

if (fstat(0, &st) < 0 || index_fd(&the_index, &oid, 0, &st, 2,
                                    prefix, 3)

It will first check if the standard input stat object is OK or not and
OR the result with index_fd() which is responsible for checking the
OK-ness of the empty tree since it checks for comparison with the tree.