Rename sha1 to committish
Rename sha1 to committish and append '--' at the end of the git
rev-parse at the end. Rename the function to
verify_submodule_committish.

Quote from Dscho:
After this, we should also append `--` to make sure that we're not parsing
this as a file name.

Two comments about naming: `sha1` is pretty misleading here, as we do not
require it to be a SHA-1 (especially in the future in which we switch to
SHA-256). Besides, what we're really asking for (via that `^0` suffix) is
a committish. Therefore, I would propose to use `committish` both in the
parameter name as well as the function name.