Try eliminating find_unique_abbrev()
Try eliminating find_uniqe_abbrev() by passing the commit hash directly
to print_submodule_summary() instead of calculating the hash using the
former.

Quote from Dscho:
That is not quite the same, as it looks for the abbreviation _in the
submodule_, not in the current project. So I think `find_unique_abbrev()`
is not correct here.

The funny thing is that we _already_ will have called `git rev-parse
--verify` for both `p->oid_src` and `p->oid_dst` in the submodule, in the
caller of this function! And while we throw away the result, and while we
do not pass `--short`, there is no reason why we shouldn't be able to do
precisely that.