Try eliminating find_unique_abbrev()
Try eliminating find_uniqe_abbrev() by passing the commit hash directly
to print_submodule_summary() instead of calculating the hash using the
former.

Quote from Dscho:
That is not quite the same, as it looks for the abbreviation _in the
submodule_, not in the current project. So I think `find_unique_abbrev()`
is not correct here.

The funny thing is that we _already_ will have called `git rev-parse
--verify` for both `p->oid_src` and `p->oid_dst` in the submodule, in the
caller of this function! And while we throw away the result, and while we
do not pass `--short`, there is no reason why we shouldn't be able to do
precisely that.

Quoting Dscho:

    It looks as if there is a whole lot of `oid_to_hex(&p->oid_src)` in that
    function. Together with the realization that we need the abbreviated
    version of that at least in one place, and the other realization that we
    already call `rev-parse --verify` for both `oid_src` and `oid_dst` in the
    caller of this function, it seems to suggest itself that we would actually
    want to pass the `--short` option, too, and to capture the output, and
    pass it down to `print_submodule_summary()` _instead of_ `missing_src` and
    `missing_dst` (e.g. as `src_abbrev` and `dst_abbrev`).
--------------------------------------------------------------------

A lot of stuff is solved but the problem is that
'verify_submodule_committish()' still returns a (null) for SHA which do
exist and are non-null. This is the reason why 13/23 tests are failing
in t7401. I also think that we need to have a "0000000" instead of a
null since that is the way the output is based for revisions which are
not existing.