Try eliminating find_unique_abbrev()
Try eliminating find_uniqe_abbrev() by passing the commit hash directly
to print_submodule_summary() instead of calculating the hash using the
former.

Quote from Dscho:
That is not quite the same, as it looks for the abbreviation _in the
submodule_, not in the current project. So I think `find_unique_abbrev()`
is not correct here.

The funny thing is that we _already_ will have called `git rev-parse
--verify` for both `p->oid_src` and `p->oid_dst` in the submodule, in the
caller of this function! And while we throw away the result, and while we
do not pass `--short`, there is no reason why we shouldn't be able to do
precisely that.

Quoting Dscho:

    It looks as if there is a whole lot of `oid_to_hex(&p->oid_src)` in that
    function. Together with the realization that we need the abbreviated
    version of that at least in one place, and the other realization that we
    already call `rev-parse --verify` for both `oid_src` and `oid_dst` in the
    caller of this function, it seems to suggest itself that we would actually
    want to pass the `--short` option, too, and to capture the output, and
    pass it down to `print_submodule_summary()` _instead of_ `missing_src` and
    `missing_dst` (e.g. as `src_abbrev` and `dst_abbrev`).

The thing is that there is some problem with
'verify_submodule_committish()' (formerly
verify_submodule_object_name()). The function isn't really giving
anything in the 'struct strbuf result'. It is just a blank. I outputted
the value of 'result.buf' and the output was nothing, literally nothing,
not a null but nothing. I am sure it is not a null value since the
commit computation is absolutely correct and it is done iff src_oid_hash
and dt_oid_hash are non-null. Therefore there is something weird going
on. I saw the 'capture_command()' usages of other (sub)commands and saw
the same usage as I am doing.

BTW I also think that we can substitute the oid_to_hex() calls in the
part where range computation happens with the src_oid_hash and
dst_oid_hash.