Simplify the 'range' computation
Simplify the computation of the 'range' variable which is used to
calculate the total_commits. The whole if-statement is very repititve
and can be shortened.

Quoting Dscho:
Since `argv_array_push()` duplicates the strings, anyway, we can totally
avoid the need for the `range` variable:

    if (IS_GITLINK(p->mod_src) && IS_GITLINK(p->mod_dst))
            argv_array_pushf(&cp_rev_list.args, "%s...%s",
			    oid_to_hex(&p->oid_src),
			    oid_to_hex(&p->oid_dst));
    else
	    argv_array_push(&cp_rev_list.args, IS_GITLINK(p->mod_src) ?
			    oid_to_hex(&p->oid_src) :
			    oid_to_hex(&p->oid_dst));

I tried his suggestion but I am not able to pass the "new range" into
the final argv_array_pushl(). I was thinking of:

argv_array_pushl(&cp_rev_list.args, "rev-list",
		"--first-parent", "--count", &cp_rev_list.argv, "--", NULL);

But this fails. What can be made better here?