Compute error message in 'generate_submodule_summary()'
Compute the error message in the 'generate_submodule_summary()'
function for more ease.

Quote from Dscho:

I am not quite sure whether it is a good idea to leave it to the
`print_submodule_summary()` function to generate the `errmsg`. I think I'd
rather have it a `char *` than an `int`.
---------------------------------------------------------------
All tests pass

Dscho's reply:
https://lore.kernel.org/git/nycvar.QRO.7.76.6.2007031712160.50@tvgsbejvaqbjf.bet/