Use get_oid() instead of spawning a new process
Use get_oid() instead of spawning a new process.
Quote from Dscho:

> +	cp_rev.git_cmd = 1;
> +	argv_array_pushl(&cp_rev.args, "rev-parse", "-q", "--verify",
> +			 argc ? argv[0] : "HEAD", NULL);

Oy. Why not simply call `get_oid()`? No need to spawn a new process.
------------------------------------------------------

What I have done is quite wrong but I cannot seem to figure out a way to
use this function apart from what Christian suggested in a previous
mail from {4:54AM (IST) 11/7/20}. Quote from Christian:

Could you tell where we are using HEAD? And if we could use something
like `get_oid("HEAD", &oid)` to get an oid that we could pass anyway?
------------------------------------------------------

After suggestions from my mentors and doing some digging on a follow-up
mail from Johannes, I figured it out. The problem I faced was that I was
using a 'struct object_id *' instead of just a 'struct object_id' due to
which the assignments and if-statement checks were not coming out to be
very proper. After looking at Dscho's mail, I realised that it will be
better to use the latter struct type since almost everywhere we use this
and the helper functions work better with this. I did some digging in
cache.h to try finding other auxiliary functions to help me. First I
tried to use the oidclr() function to clear the object_id to NULL in
case of when a '--files' is passed but that wasn't the right function to
use (something I realised after concentrating on the definition of the
function really hard).

So here we are, with the fix. All tests pass.