Use is_submodule_active() instead of spawning a process
Use is_submodule_active() instead of spawning a new process for every
submodule in prepare_submodule_summary(). Quote from Dscho:

> +		/* Also show added or modified modules which are checked out */
> +		cp_rev_parse.dir = p->sm_path;
> +		cp_rev_parse.git_cmd = 1;
> +		cp_rev_parse.no_stderr = 1;
> +		cp_rev_parse.no_stdout = 1;
> +
> +		argv_array_pushl(&cp_rev_parse.args, "rev-parse",
> +				 "--git-dir", NULL);
> +
> +		if (!run_command(&cp_rev_parse))

I wonder whether we really need to waste an entire spawned process on
figuring out whether `p->sm_path` refers to an active repository. Wouldn't
`is_submodule_active(r, p->sm_path)` fulfill the same purpose?
-----------------------------------------------------------

I have currently made a makeshift arrangement to make this work. The
thing is that the function is_submodule_active() actually tries to
access the .gitmodules to check if the submodule is active or not. But
the thing is that in case of the 'for-status' option, a .gitmodules is
not accessible since we "hide" it by sparse-checking out, deleting or
other creative ways we may desire. Therefore, the test fails in case of
the 'for-status' option, but mind you that this happens only for the
'for-status' test of t7418 and not the one in t7401 because in the
latter we never really deleted or "hid" the .gitmodules file and hence
the test passes in t7401.

The arrangement now is that the is_submodule_active() is called for all
cases except 'for-status'. In case of 'for-status' we spawn a CP. This
has some improvement over the original way this arrangement worked by
spwaning a CP irrespective of the situation.

This needs some improvement for the case of 'for-status'. Maybe we can
improve is_submodule_active() to fetch activity status of a submodule
from a previous revision of .gitmodules which we may have.

https://lore.kernel.org/git/nycvar.QRO.7.76.6.2007031712160.50@tvgsbejvaqbjf.bet/