Use is_submodule_active() instead of spawning a process
Use is_submodule_active() instead of spawning a new process for every
submodule in prepare_submodule_summary(). Quote from Dscho:

> +		/* Also show added or modified modules which are checked out */
> +		cp_rev_parse.dir = p->sm_path;
> +		cp_rev_parse.git_cmd = 1;
> +		cp_rev_parse.no_stderr = 1;
> +		cp_rev_parse.no_stdout = 1;
> +
> +		argv_array_pushl(&cp_rev_parse.args, "rev-parse",
> +				 "--git-dir", NULL);
> +
> +		if (!run_command(&cp_rev_parse))

I wonder whether we really need to waste an entire spawned process on
figuring out whether `p->sm_path` refers to an active repository. Wouldn't
`is_submodule_active(r, p->sm_path)` fulfill the same purpose?
-----------------------------------------------------------

t7418 fails since it has the check for 'for-status'.
is_submodule_active() accesses the .gitmodules for checking the activity
of the submodule and since this is missing for the 'for-status', we get
a failure. t7401 also has a test for 'for-status' but since we never
really delete/hide the .gitmodules in by any means in that case, we
never really get a failure. t7508 passes.

UPDATE: I tried this makeshift mechanism. This is no way near perfect
but it runs a different set of instructions for the case when
'for-status' is 1, therefore creating a new process. I know this is not
the best solution but atleast we only spawn a process when for-status is
0 so I guess that is a small step?

https://lore.kernel.org/git/nycvar.QRO.7.76.6.2007031712160.50@tvgsbejvaqbjf.bet/