Fix the 'log' issue
Fix the issue where we were getting an extra 'git log' failure output
for submodules which did not really exist in the working directory of
the superproject. The thing was to run a 'git log' only for cases where
we had a commit tally to show the commits of.
-----8<-----
Output before:
$ git submodule summary 4efb51bdef8916013b8490ef087a1c3712b907c7 -- lxconf-bash
fatal: exec 'rev-parse': cd to 'lxconf-bash' failed: No such file or directory
* lxconf-bash 2c819db...0000000:
fatal: exec 'log': cd to 'lxconf-bash' failed: No such file or directory
-----8<-----
Output now:
$ git submodule summary 4efb51bdef8916013b8490ef087a1c3712b907c7 -- lxconf-bash
fatal: not a git repository: 'lxconf-bash/.git'
* lxconf-bash 2c819db...0000000:
-----8<-----